Note for future progress
We can think about making BoggleGUI a child class for BoggleGame, and creating an instance of the game everytime a game starts. But if we do that, we should make the board and other global variables into instance variables. We would also need to change the inheritance structure, since we wouldn't be able to extend from JFrame here